backend/requirements.txt
asgiref==3.7.2
boto3==1.17.58
botocore==1.20.112
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.47.0
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-filter==21.1
django-rest-auth-forked==0.9.5
django-rest-passwordreset==1.2.1
django-sendgrid-v5==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pdfkit==0.6.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.4.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
